parsing response funtion + http get request
parsing response funtion + http get request + adjusting stlying and formats of the entries

+++ Identifyer_App/app/src/main/java/com/example/identifyer/MainActivity.kt
@@ -1,9 +1,13 @@
 package com.example.identifyer
 
+
 import android.Manifest
+import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Bundle
+import android.os.StrictMode
+import android.os.StrictMode.ThreadPolicy
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
@@ -16,18 +20,25 @@ import androidx.lifecycle.lifecycleScope
 import com.budiyev.android.codescanner.*
 import com.example.identifyer.ViewModel.InmateViewModel
 import com.example.identifyer.ViewModel.RoomViewModel
-import com.example.identifyer.model.Room
 import com.example.identifyer.ViewModel.UserViewModel
-
 import com.example.identifyer.model.Inmate
-import com.example.identifyer.model.User
+import com.example.identifyer.model.Room
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-
 import kotlinx.coroutines.*
-import kotlinx.serialization.json.JsonObject
 import org.json.JSONObject
+import org.xmlpull.v1.XmlPullParserException
+import java.io.BufferedReader
+import java.io.IOException
+import java.lang.Exception
+import java.net.HttpURLConnection
+import java.net.MalformedURLException
+import java.net.URL
+import java.text.SimpleDateFormat
+import java.util.*
+import java.util.concurrent.TimeUnit
 
+    //Main
 class MainActivity : AppCompatActivity() {
 
 
@@ -42,19 +53,11 @@ class MainActivity : AppCompatActivity() {
     //View Model user
     lateinit var userViewModel:UserViewModel
 
-    //Testdata
-    var jsonTestData : String = "{\n" +
-            "  \"room\": [\n" +
-            "    {\"id\":\"1\", \"roomNumber\":\"1.11\", \"tract\" : \"A\", \"securityLevel\":\"1\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}, \n" +
-            "  {\"id\":\"2\", \"roomNumber\":\"2.22\", \"tract\" : \"B\", \"securityLevel\":\"2\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}, \n" +
-            "  {\"id\":\"3\", \"roomNumber\":\"3.33\", \"tract\" : \"C\", \"securityLevel\":\"3\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}, \n" +
-            "  {\"id\":\"4\", \"roomNumber\":\"4.44\", \"tract\" : \"D\", \"securityLevel\":\"4\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}, \n" +
-            "  {\"id\":\"5\", \"roomNumber\":\"5.55\", \"tract\" : \"E\", \"securityLevel\":\"5\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}, \n" +
-            "  {\"id\":\"6\", \"roomNumber\":\"6.66\", \"tract\" : \"F\", \"securityLevel\":\"6\", \"maxCapacity\": \"4\", \"inmateList\":[\"1\",\"3\", \"6\"]}], \n" +
-            "  \"inmates\": [\n" +
-            "    {\"id\": \"1\", \"firstname\" : \"John\", \"lastname\": \"Doe\", \"gender\": \"male\", \"dateOfBirth\": \"11.11.1911\", \"nationality\":\"Austria\", \"sentence\": \"Robbery\", \"arrivalDate\": \"23.4.2020\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"1\", \"physicalWellness\":\"No Comment\", \"mentalWellness\":\"Depressions\", \"combatExperience\":\"7 years of UFC, Gangmember\", \"additionalNotes\":[\"Former Gangmember\", \"Likes to be funny\", \"sarcastic\"], \"room_id\": \"1\"}, {\"id\": \"2\", \"firstname\" : \"Nils\", \"lastname\": \"Petsch\", \"gender\": \"male\", \"dateOfBirth\": \"24.05.1999\", \"nationality\":\"Austria\", \"sentence\": \"Robbery\", \"arrivalDate\": \"25.05.2021\", \"TimeOfSentence\":\"5 Years\",\"securityLevel\":\"4\", \"physicalWellness\":\"knee porblems\", \"mentalWellness\":\"No comment\", \"combatExperience\":\"7 Karate, 3 Years Muy Thai\", \"additionalNotes\":[\"Former Footballer\", \"Likes to sniff gas\", \"loves alcohol\"], \"room_id\": \"2\"}, {\"id\": \"3\", \"firstname\" : \"Flo\", \"lastname\": \"Huber\", \"gender\": \"male\", \"dateOfBirth\": \"24.07.1988\", \"nationality\":\"Austria\", \"sentence\": \"Played Yi in League of Legends\", \"arrivalDate\": \"15.03.2021\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"3\", \"physicalWellness\":\"No Comment\", \"mentalWellness\":\"Who the fuck plays YI ???\", \"combatExperience\":\"One punch and u gone\", \"additionalNotes\":[\"Former E-Sport Player\", \"Likes small things\", \"loves YI\"], \"room_id\": \"3\"}, {\"id\": \"4\", \"firstname\" : \"Felix\", \"lastname\": \"Engelmeier\", \"gender\": \"male\", \"dateOfBirth\": \"14.12.1999\", \"nationality\":\"Austria\", \"sentence\": \"Did the homework\", \"arrivalDate\": \"10.06.2022\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"2\", \"physicalWellness\":\"No Comment\", \"mentalWellness\":\"No comment\", \"combatExperience\":\"10 Years Muy Thai\", \"additionalNotes\":[\"Former Pro Chess Player\", \"Likes to sniff on glue\", \"loves alcohol\"], \"room_id\": \"4\"}, {\"id\": \"5\", \"firstname\" : \"Lara\", \"lastname\": \"Roth\", \"gender\": \"female\", \"dateOfBirth\": \"13.11.1998\", \"nationality\":\"Austria\", \"sentence\": \"Supporting Planktion\", \"arrivalDate\": \"25.10.2021\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"4\", \"physicalWellness\":\"trained \", \"mentalWellness\":\"No comment\", \"combatExperience\":\"7 Karate, 3 Years Muy Thai\", \"additionalNotes\":[\"Former Athlete\", \"Likes to sry\", \"loves alcohol\"], \"room_id\": \"5\"}, {\"id\": \"6\", \"firstname\" : \"Patrick\", \"lastname\": \"Star\", \"gender\": \"male\", \"dateOfBirth\": \"13.12.1998\", \"nationality\":\"Austria\", \"sentence\": \"Eating a burger at 3 am\", \"arrivalDate\": \"25.10.2021\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"4\", \"physicalWellness\":\"star fish \", \"mentalWellness\":\"dumb\", \"combatExperience\":\"weights a lot\", \"additionalNotes\":[\"Former Sum-Ringer\", \"Likes to eat\", \"loves spongebob\"], \"room_id\": \"6\"}, {\"id\": \"7\", \"firstname\" : \"Spongebob\", \"lastname\": \"Schwammkopf\", \"gender\": \"male\", \"dateOfBirth\": \"23.12.1998\", \"nationality\":\"Austria\", \"sentence\": \"Forgetting how to do a crabby burger\", \"arrivalDate\": \"25.10.2021\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"5\", \"physicalWellness\":\"sponge\", \"mentalWellness\":\"smart sponge\", \"combatExperience\":\"spacuala master\", \"additionalNotes\":[\"Former Cook\", \"Likes to cook\", \"loves patrick\"], \"room_id\": \"1\"},{\"id\": \"8\", \"firstname\" : \"Tadeus\", \"lastname\": \"Tentakel\", \"gender\": \"male\", \"dateOfBirth\": \"23.12.1998\", \"nationality\":\"Austria\", \"sentence\": \"Forgetting how to do a crabby burger\", \"arrivalDate\": \"25.10.2021\", \"timeOfSentence\":\"5 Years\",\"securityLevel\":\"5\", \"physicalWellness\":\"sponge\", \"mentalWellness\":\"smart sponge\", \"combatExperience\":\"spacuala master\", \"additionalNotes\":[\"Former Cook\", \"Likes to cook\", \"loves patrick\"], \"room_id\": \"1\"}\n" +
-            "  ]\n" +
-            "}";
+    //ngrok URL which redirect to the port where the server runs --> Flo: 8080, Nils: 8081
+    //url needs to be updated when new instance of--> ngrok http 8080 (port of the backend)
+    //ngrok http 8080 --> creates and url which points at the pc port | example: https://fbaa-82-218-247-87.eu.ngrok.io -> points at -> http://localhost:8080/ (when backend runs on port: 8080)
+    private val url :String = "https://fbaa-82-218-247-87.eu.ngrok.io"
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -64,6 +67,8 @@ class MainActivity : AppCompatActivity() {
         inmateViewModel = ViewModelProvider(this).get(InmateViewModel::class.java)
         userViewModel = ViewModelProvider(this).get(UserViewModel::class.java)
 
+        val policy = ThreadPolicy.Builder().permitAll().build()
+        StrictMode.setThreadPolicy(policy)
 
         if( ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_DENIED){
             ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.CAMERA),123)
@@ -73,36 +78,66 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    //deleting rooms and inmates
-     suspend private fun deleteAllRoomsInmates(){
 
-        roomViewModel.deleteAllRooms();
-        inmateViewModel.deleteAllInmates();
+    //deletes all rooms and inmates
+    private suspend fun deleteAllRoomsInmates(){
+        roomViewModel.deleteAllRooms()
+        inmateViewModel.deleteAllInmates()
     }
-        //response from http request --> parses Response and inserts the data
-    private fun parseJson(reponse:String){
-
-        var obj = JSONObject(reponse);
-        var roomObj = obj["room"].toString();
-        var inmateObj = obj["inmates"].toString();
 
 
-        val typeTokenRooms = object : TypeToken<List<Room>>(){}.type;
-        val rooms = Gson().fromJson<List<Room>>(roomObj, typeTokenRooms);
+    private fun parseJson(response:String){
 
-        val typeTokenInmates = object : TypeToken<List<Inmate>>(){}.type;
-        val inmates = Gson().fromJson<List<Inmate>>(inmateObj, typeTokenInmates);
+        val obj = JSONObject(response)
 
-        for (item: Room in rooms ){
-            val newRoom : Room = Room(item.id,item.roomNumber, item.inmateList, item.securityLevel, item.tract, item.maxCapacity );
+        val roomsObj = obj["room"].toString()
+        val inmatesObj  =obj["inmates"].toString()
+
+        val typeTokenRooms = object : TypeToken<List<ResponseRoom>>(){}.type
+        val rooms = Gson().fromJson<List<ResponseRoom>>(roomsObj, typeTokenRooms)
+
+        val typeTokenInmates = object : TypeToken<List<ResponseInmate>>(){}.type
+        val inmates = Gson().fromJson<List<ResponseInmate>>(inmatesObj, typeTokenInmates)
+
+        for (item: ResponseRoom in rooms){
+            val newRoom : Room = Room(item.id,item.roomNumber, item.securityLevel, item.tractName, item.maxInmateCount )
             roomViewModel.insert(newRoom)
         }
 
-        for (item: Inmate in inmates){
-            val newInmate : Inmate = Inmate(item.id, item.firstname, item.lastname, item.gender, item.dateOfBirth, item.nationality, item.sentence, item.arrivalDate, item.timeOfSentence,item.securityLevel, item.physicalWellness, item.mentalWellness, item.combatExperience, item.additionalNotes, item.room_id)
+        for (item: ResponseInmate in inmates){
+            val offenseList = ArrayList<String>();
+            for(entry: Offense in item.offenseList){
+                val s = "# "+ entry.id + " - " + entry.description + "\n" + convertLongToTime(entry.timeOfOffense)
+                offenseList.add(s)
+            }
+
+            var nationality = item.nationality.joinToString(", ")
+
+            var securityLevel = when(item.securityLevel){
+                "LEVEL_HIGH"-> "High"
+                "LEVEL_MEDIUM"-> "Medium"
+                "LEVEL_LOW"-> "Low"
+                else-> {
+                    "NONE"
+                }
+            }
+
+            var gender = when(item.gender){
+                "GENDER_MALE"-> "male"
+                "GENDER_FEMALE"-> "female"
+                "GENDER_OTHER"-> "other"
+                else-> {
+                    "NONE"
+                }
+            }
+            val newInmate : Inmate = Inmate(item.id, item.firstName, item.lastName, gender, convertLongToTimeWithoutTime(item.dateOfBirth), nationality, (item.sentence/31536000).toString() + " years", convertLongToTimeWithoutTime(item.arrivalDate), (TimeUnit.DAYS.convert(item.sentence, TimeUnit.NANOSECONDS)/30).toString(),securityLevel, item.physicalWellness, item.mentalWellness,offenseList, item.combatExperience.joinToString("\n#"), item.notes, item.room.id)
             inmateViewModel.insert(newInmate)
+
         }
+
     }
+
+
 
     //Scanning function
     private fun startScanning(){
@@ -112,7 +147,7 @@ class MainActivity : AppCompatActivity() {
         codeScanner.formats = CodeScanner.ALL_FORMATS
         codeScanner.autoFocusMode = AutoFocusMode.SAFE
         codeScanner.scanMode = ScanMode.SINGLE
-        codeScanner.isAutoFocusEnabled= true
+        codeScanner.isAutoFocusEnabled = true
         codeScanner.isFlashEnabled= false
 
         codeScanner.decodeCallback = DecodeCallback{
@@ -120,7 +155,7 @@ class MainActivity : AppCompatActivity() {
             val textString = it.text.toString()
 
             runOnUiThread {
-                Toast.makeText(this, "Roomnumber :${it.text.toString()} ", Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, "RoomName :${it.text.toString()} ", Toast.LENGTH_SHORT).show()
             }
 
 
@@ -171,7 +206,6 @@ class MainActivity : AppCompatActivity() {
             codeScanner?.releaseResources()
         }
         super.onPause()
-
     }
 
 
@@ -193,16 +227,80 @@ class MainActivity : AppCompatActivity() {
             //sync Database
             syncDatabase()
             return true
+        }else if(itemId == R.id.action_help){
+            val intent = Intent(this, HelpActivity::class.java)
+            startActivity(intent)
         }
         return super.onOptionsItemSelected(item)
     }
     //method which calls the http request method --> gets response as String which-->validate if empty --> delete all entries if not empty--> parses response and inserts the data
     private fun syncDatabase(){
+           var response = requestData(url)
         lifecycleScope.launch {
-            deleteAllRoomsInmates()
-            delay(1000)
-            parseJson(jsonTestData)
+            if(!response.isNullOrBlank()){
+                deleteAllRoomsInmates()
+                delay(1000)
+                try {
+                    parseJson(response)
+                    delay(1000)
+                    Toast.makeText(applicationContext,"Database updated successfully!", Toast.LENGTH_SHORT).show()
+                }catch (e:Exception){
+                    Toast.makeText(applicationContext,"Database update failed!", Toast.LENGTH_SHORT).show()
+                }
+            }else{
+                Toast.makeText(applicationContext,"No Data found!", Toast.LENGTH_SHORT).show()
+            }
         }
+    }
+    //http get call to a an address which redirects to the server localhost 8081
+    // (command in ngronk) : ngronk http 8081 --> creates url which is visible for everyone!  points at localhost:8081 from the server (command giving device)
+    private fun requestData(url:String) :String{
+
+        var data: String =""
+        var con: HttpURLConnection? = null
+        try {
+
+            val url = URL(url)
+            con = url.openConnection() as HttpURLConnection
+            con.requestMethod = "GET"
+
+            Log.d("response", con.inputStream.toString())
+            val reader = BufferedReader(con.inputStream.reader())
+
+            try {
+                data = reader.readText()
+            } finally {
+                reader.close()
+            }
+        } catch (ex: MalformedURLException) {
+            Log.e("LOG_TAG", ex.toString())
+
+
+        } catch (ex: XmlPullParserException) {
+            Log.e("LOG_TAG", ex.toString())
+
+
+        } catch (ex: IOException) {
+            Log.e("LOG_TAG", ex.toString())
+
+        } finally {
+            con?.disconnect()
+        }
+
+        return data
+    }
+
+    //converts Long to Date Format
+    private fun convertLongToTime(time: Long): String {
+        val date = Date(time)
+        val format = SimpleDateFormat("yyyy.MM.dd HH:mm")
+        return format.format(date)
+    }
+    //converts Long to Date Format with out day time
+    private fun convertLongToTimeWithoutTime(time: Long): String {
+        val date = Date(time)
+        val format = SimpleDateFormat("yyyy.MM.dd")
+        return format.format(date)
     }
 
 }
